consolidate c2p configfile parser in each pvp
