+++ pkg/kyverno/oscal2policy_test.go
@@ -21,7 +21,6 @@ import (
 	"testing"
 
 	"github.com/IBM/compliance-to-policy/pkg"
-	"github.com/IBM/compliance-to-policy/pkg/c2pcr"
 	typec2pcr "github.com/IBM/compliance-to-policy/pkg/types/c2pcr"
 	"github.com/stretchr/testify/assert"
 )
@@ -71,7 +70,7 @@ func TestOscal2Policy(t *testing.T) {
 			Namespace: "",
 		},
 	}
-	c2pcrParser := c2pcr.NewParser(gitUtils)
+	c2pcrParser := NewParser(gitUtils)
 	c2pcrParsed, err := c2pcrParser.Parse(c2pcrSpec)
 	assert.NoError(t, err, "Should not happen")
